+++ app/controllers/analytics/group_controller.rb
@@ -10,7 +10,7 @@ module Analytics
     # rubocop:disable Metrics/MethodLength
     def performance_per_group
       groups_lesson_summaries = if selected_param_present_but_not_all?(@selected_group_id)
-                                  GroupLessonSummary.joins(:group).where(group_id: @selected_group_id)
+                                  GroupLessonSummary.joins(:group).where(group_id: @selected_group_id, groups: { deleted_at: nil })
                                 elsif selected_param_present_but_not_all?(@selected_chapter_id)
                                   GroupLessonSummary.joins(:group).where(groups: { deleted_at: nil }).where(chapter_id: @selected_chapter_id)
                                 else