+++ spec/controllers/lessons_controller_spec.rb
@@ -71,6 +71,19 @@ RSpec.describe LessonsController, type: :controller do
       end
     end
 
+    describe '#edit' do
+      before :each do
+        @lesson = create :lesson
+        get :edit, params: { id: @lesson.id }
+      end
+
+      it { should respond_with 200 }
+      it { should render_template 'edit' }
+      it 'presents a lesson to edit' do
+        expect(assigns(:lesson)).to eq @lesson
+      end
+    end
+
     describe '#create' do
       context 'Creates the lesson successfully' do
         before :each do
@@ -94,6 +107,7 @@ RSpec.describe LessonsController, type: :controller do
           expect(created_lesson.subject).to eq @subject
         end
       end
+
       context 'Tries to create an already existing lesson' do
         before :each do
           @group = create :group
@@ -108,10 +122,54 @@ RSpec.describe LessonsController, type: :controller do
           } }
         end
 
-        it do
-          should render_template :new
-          should respond_with :unprocessable_entity
-          should set_flash[:failure_notice]
+        it { should render_template :new }
+        it { should respond_with :unprocessable_entity }
+        it { should set_flash[:failure_notice] }
+      end
+
+      context 'Tries to create a lesson with missing fields' do
+        before :each do
+          post :create, params: { lesson: {
+            group_id: nil,
+            subject_id: nil,
+            date: ''
+          } }
+        end
+
+        it { should render_template :new }
+        it { should respond_with :unprocessable_entity }
+        it { should set_flash[:failure_notice] }
+      end
+    end
+
+    describe '#update' do
+      before :each do
+        @lesson = create :lesson
+      end
+
+      context 'success' do
+        before :each do
+          post :update, params: { id: @lesson.id, lesson: { group_id: @lesson.group_id, subject_id: @lesson.subject_id, date: 1.day.from_now } }
+        end
+
+        it { should respond_with 302 }
+        it { should redirect_to lesson_url }
+        it { should set_flash[:success_notice] }
+
+        it 'updates the lesson' do
+          expect(@lesson.reload.date).to eq 1.day.from_now.to_date
+        end
+      end
+
+      context 'failure' do
+        before :each do
+          post :update, params: { id: @lesson.id, lesson: { group_id: @lesson.group_id, subject_id: @lesson.subject_id, date: nil } }
+        end
+
+        it { should respond_with 422 }
+        it { should render_template :edit }
+        it 'does not update the lesson' do
+          expect(@lesson.reload.date).not_to eq 1.day.from_now.to_date
         end
       end
     end