+++ app/components/common_components/tooltip_component.rb
@@ -1,5 +1,6 @@
 class CommonComponents::TooltipComponent < ViewComponent::Base
-  def initialize(text:)
+  def initialize(text:, custom_class: nil)
     @text = text
+    @custom_class = custom_class
   end
 end