+++ app/components/group_enrolled_students_component.rb
@@ -6,7 +6,8 @@ class GroupEnrolledStudentsComponent < ViewComponent::Base
       <% card.with_card_content do %>
         <%= render StudentTableForm.new(student: @new_student, group: @group, is_edit: false) if helpers.policy(@new_student).create? %>
         <% if @group.students.any? %>
-          <%= render TableComponents::Table.new(pagy: @pagy, options: { no_pagination: true, turbo_id: 'students' }, rows: @students, row_component: TableComponents::StudentTurboRow, row_arguments: { group: @group }) %>
+          <%= render TableComponents::Table.new(pagy: @pagy, options: { no_pagination: true, turbo_id: 'students' }, rows: @students, row_component: TableComponents::StudentTurboRow,
+           row_arguments: { group: @group, students_with_invalid_grades: @students_with_invalid_grades }) %>
         <% else %>
           <span class="w-full p-2 text-center bg-gray-50 border-b border-gray-200 text-md font-medium text-gray-500 uppercase tracking-wider">No students enrolled yet</span>
         <% end %>
@@ -14,8 +15,9 @@ class GroupEnrolledStudentsComponent < ViewComponent::Base
     <% end %>
   ERB
 
-  def initialize(students:, group:)
+  def initialize(students:, group:, students_with_invalid_grades: nil)
     @student_records = students.where(deleted_at: nil)
+    @students_with_invalid_grades = students_with_invalid_grades
     @group = group
     @new_student = Student.new
     @new_student.enrollments.build(group: @group)