+++ app/services/csv_service.rb
@@ -2,6 +2,8 @@ class CsvService
   class << self
     require 'csv'
 
+    # rubocop:disable Metrics/MethodLength
+    # rubocop:disable Metrics/AbcSize
     def deserialize_students(file)
       content = File.open(file, encoding: 'UTF-8', &:read)
       content.sub!("\xEF\xBB\xBF", '')
@@ -10,9 +12,40 @@ class CsvService
         {
           first_name: row['First Name'],
           last_name: row['Last Name'],
-          gender: row['Gender'],
-          dob: safe_parse_date(row['Date of Birth'])
+          gender: safe_parse_gender(row['Gender']),
+          dob: safe_parse_date(row['Date of Birth']),
+          estimated_dob: safe_parse_bool?(row['DOB Estimated']),
+          country_of_nationality: row['Country of Nationality'],
+          family_members: row['Family Members'],
+          guardian_name: row['Guardian Name'],
+          guardian_contact: row['Guardian Contact'],
+          guardian_occupation: row['Guardian Occupation'],
+          health_insurance: row['Health Insurance'],
+          health_issues: row['Health Issues'],
+          hiv_tested: safe_parse_bool?(row['Hiv Tested']),
+          name_of_school: row['Name of School'],
+          notes: row['Notes'],
+          reason_for_leaving: row['Reason for Leaving'],
+          school_level_completed: row['School Level Completed'],
+          year_of_dropout: safe_parse_integer(row['Year of Dropout'])
         }
+      end
+    end
+    # rubocop:enable Metrics/MethodLength
+    # rubocop:enable Metrics/AbcSize
+
+    def safe_parse_integer(value)
+      Integer(value) if value
+    rescue ArgumentError
+      nil
+    end
+
+    def safe_parse_bool?(value)
+      case value.to_s.downcase
+      when 'true', 't', 'yes', 'y', '1', 'True'
+        true
+      else
+        false
       end
     end
 
@@ -20,6 +53,19 @@ class CsvService
       date_text.present? ? Date.parse(date_text) : Time.zone.today
     rescue Date::Error
       Time.zone.today
+    end
+
+    def safe_parse_gender(gender)
+      return gender if %w[M F NB].include? gender
+
+      case gender[0]
+      when 'M'
+        :M
+      when 'F'
+        :F
+      else
+        :NB
+      end
     end
   end
 end