+++ spec/policies/tag_policy_spec.rb
@@ -208,12 +208,15 @@ RSpec.describe TagPolicy do
       let(:current_user) { create :admin_of, organization: org1 }
       subject(:result) { TagPolicy::Scope.new(current_user, Tag).resolve_for_organization_id(org1.id) }
 
-      it 'scopes to organization and shared tags' do
+      it 'scopes to organization and shared tags and tags with shared organization ids' do
         shared_tags = create_list :tag, 3
         org1_tag = create :tag, organization: org1, shared: false
         org2_tag = create :tag, organization: org2, shared: false
+        shared_org2_tag = create :tag, organization: org2, shared: true, shared_organization_ids: [org1.id]
+
         expect(result).to include(shared_tags[0], shared_tags[1], shared_tags[2])
         expect(result).to include(org1_tag)
+        expect(result).to include(shared_org2_tag)
         expect(result).not_to include(org2_tag)
       end
     end
@@ -224,13 +227,14 @@ RSpec.describe TagPolicy do
       let(:shared_tags) { create_list :tag, 3 }
       let(:org1_tag) { create :tag, organization: org1, shared: false }
       let(:org2_tag) { create :tag, organization: org2, shared: false }
+      let(:shared_org2_tag) { create :tag, organization: org2, shared: false, shared_organization_ids: [org1.id] }
       subject(:result) { TagPolicy::Scope.new(current_user, Tag).resolve }
 
       context 'user is a super administrator' do
         let(:current_user) { create :super_admin }
 
         it 'includes all tags' do
-          expect(result).to include shared_tags[0], shared_tags[1], shared_tags[2], org1_tag, org2_tag
+          expect(result).to include shared_tags[0], shared_tags[1], shared_tags[2], org1_tag, org2_tag, shared_org2_tag
         end
       end
 
@@ -238,7 +242,7 @@ RSpec.describe TagPolicy do
         let(:current_user) { create :admin_of, organization: org1 }
 
         it 'includes all tags' do
-          expect(result).to include shared_tags[0], shared_tags[1], shared_tags[2], org1_tag
+          expect(result).to include shared_tags[0], shared_tags[1], shared_tags[2], org1_tag, shared_org2_tag
           expect(result).not_to include org2_tag
         end
       end