+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Added ability to update a lesson's date
+- Fixed Organization and Chapter summaries showing incorrect counts
 
 ## 0.38.1 - Fix Student Analytics Summaries
 - Fixed Student Analytics Summaries causing crash when viewing Analytics