+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
+- Added ability to update a lesson's date
 - Added ability to filter by single group in 'Group analytics'
+- Fixed Organization and Chapter summaries showing incorrect counts
 
 ## 0.38.1 - Fix Student Analytics Summaries
 - Fixed Student Analytics Summaries causing crash when viewing Analytics