+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Extended importing students to handle all student fields
 - Added alert when viewing a group with students graded outside their enrollment
 - Improved speed when querying students which have no active enrollments 
 - Removed 'Created' and added 'Enrolled Since' column to student table in group view